final edit

+++ SpotBot/bot.py
@@ -1,7 +1,5 @@
 import asyncio
 import urllib.request
-
-from telegram import MessageId
 from SpotBot.spotify import MusicDownloader, AlbumDownloader
 from pyrogram.types.bots_and_keyboards.reply_keyboard_markup import ReplyKeyboardMarkup
 from pyrogram.types.bots_and_keyboards.inline_keyboard_button import InlineKeyboardButton
@@ -50,17 +48,18 @@ async def mp3dl(Client, msg: Message):
         inputs = await Client.ask(chat_id=msg.from_user.id, text='Paste Spotify song url or enter search query\nExample: **Cool - Dua Lipa**\nAbort /cancel', reply_markup=ReplyKeyboardRemove(True))
         if not inputs.text == '/cancel':
             info_searching = await icbot.send_message(chat_id=msg.chat.id, text='__Searching for result.__')
-            spotify.Finder(0, inputs.text)
-            await icbot.edit_message_text(chat_id=msg.chat.id, text=f'**Found** : {spotify.Sname}', message_id=info_searching.id)
-            await asyncio.sleep(0.5)
-            await icbot.edit_message_text(
-                chat_id=msg.chat.id, text=f'**Downloading** {spotify.Sname}', message_id=info_searching.id)
-            await MusicDownloader(
-                query=inputs.text, format='mp3')
-            await icbot.delete_messages(msg.chat.id, info_searching.id)
-            await msg.reply_text(f'{spotify.Scaption}\n\n**Download of {spotify.Filename} complete.**\n__Uploading to you ...__', reply_markup=InlineKeyboardMarkup([[InlineKeyboardButton('Album Cover', url=spotify.Scover)]]))
-            await msg.reply_audio(audio=spotify.Filename, caption=f'🎧 Downloaded By [{Configs.BOT_NAME}](https://telegram.me/{Configs.BOT_UNAME}) Bot', reply_markup=MAIN_KEYS)
-            os.remove(spotify.Filename)
+            result = spotify.Finder(0, inputs.text)
+            await icbot.edit_message_text(chat_id=msg.chat.id, text=spotify.Sname, message_id=info_searching.id)
+            if result:
+                await asyncio.sleep(0.5)
+                await icbot.edit_message_text(
+                    chat_id=msg.chat.id, text=f'**Downloading** {spotify.Sname}', message_id=info_searching.id)
+                await MusicDownloader(
+                    query=inputs.text, format='mp3')
+                await icbot.delete_messages(msg.chat.id, info_searching.id)
+                await msg.reply_text(f'{spotify.Scaption}\n\n**Download of {spotify.Filename} complete.**\n__Uploading to you ...__', reply_markup=InlineKeyboardMarkup([[InlineKeyboardButton('Album Cover', url=spotify.Scover)]]))
+                await msg.reply_audio(audio=spotify.Filename, caption=f'🎧 Downloaded By [{Configs.BOT_NAME}](https://telegram.me/{Configs.BOT_UNAME}) Bot', reply_markup=MAIN_KEYS)
+                os.remove(spotify.Filename)
         else:
             await msg.reply_text(f'You\'ve backed to main.', reply_markup=MAIN_KEYS)
 
@@ -80,20 +79,21 @@ async def flacdl(Client, msg: Message):
         inputs = await Client.ask(chat_id=msg.from_user.id, text='FLAC Downloader\nPaste Spotify song url or enter search query\nExample: **Cool - Dua Lipa**\nAbort /cancel', reply_markup=ReplyKeyboardRemove(True))
         if not inputs.text == '/cancel':
             info_searching = await icbot.send_message(chat_id=msg.chat.id, text='__Searching for result.__')
-            spotify.Finder(0, inputs.text)
-            await icbot.edit_message_text(chat_id=msg.chat.id, text=f'**Found** : {spotify.Sname}', message_id=info_searching.id)
-            await asyncio.sleep(0.5)
-            await icbot.edit_message_text(
-                chat_id=msg.chat.id, text=f'**Downloading** {spotify.Sname}', message_id=info_searching.id)
-            await MusicDownloader(
-                query=inputs.text, format='flac')
-            await icbot.delete_messages(msg.chat.id, info_searching.id)
-            await msg.reply_text(f'{spotify.Scaption}\n\n**Download of {spotify.Filename} complete.**\n__Uploading to you ...__', reply_markup=InlineKeyboardMarkup([[InlineKeyboardButton('Album Cover', url=spotify.Scover)]]))
-            urllib.request.urlretrieve(spotify.Scover, 'thumb.jpg')
-            await msg.reply_audio(audio=spotify.Filename, caption=f'🎧 Downloaded By [{Configs.BOT_NAME}](https://telegram.me/{Configs.BOT_UNAME}) Bot', thumb='thumb.jpg', performer=spotify.Sartist, reply_markup=MAIN_KEYS)
-            os.remove(spotify.Filename)
-            os.remove('thumb.jpg')
-            os.remove(spotify.IFilename)
+            result = spotify.Finder(0, inputs.text)
+            await icbot.edit_message_text(chat_id=msg.chat.id, text=spotify.Sname, message_id=info_searching.id)
+            if result:
+                await asyncio.sleep(0.5)
+                await icbot.edit_message_text(
+                    chat_id=msg.chat.id, text=f'**Downloading** {spotify.Sname}', message_id=info_searching.id)
+                await MusicDownloader(
+                    query=inputs.text, format='flac')
+                await icbot.delete_messages(msg.chat.id, info_searching.id)
+                await msg.reply_text(f'{spotify.Scaption}\n\n**Download of {spotify.Filename} complete.**\n__Uploading to you ...__', reply_markup=InlineKeyboardMarkup([[InlineKeyboardButton('Album Cover', url=spotify.Scover)]]))
+                urllib.request.urlretrieve(spotify.Scover, 'thumb.jpg')
+                await msg.reply_audio(audio=spotify.Filename, caption=f'🎧 Downloaded By [{Configs.BOT_NAME}](https://telegram.me/{Configs.BOT_UNAME}) Bot', thumb='thumb.jpg', performer=spotify.Sartist, reply_markup=MAIN_KEYS)
+                os.remove(spotify.Filename)
+                os.remove('thumb.jpg')
+                os.remove(spotify.IFilename)
         else:
             await msg.reply_text(f'You\'ve backed to main.', reply_markup=MAIN_KEYS)
 
@@ -113,16 +113,17 @@ async def album(Client, msg: Message):
         inputs = await Client.ask(chat_id=msg.from_user.id, text='Album Downloader\nPaste Spotify album url\nAbort /cancel', reply_markup=ReplyKeyboardRemove(True))
         if not inputs.text == '/cancel':
             info_searching = await icbot.send_message(chat_id=msg.chat.id, text='__Searching for result.__')
-            spotify.Finder(0, inputs.text)
-            await icbot.edit_message_text(chat_id=msg.chat.id, text=f'**Found** : {spotify.Sname}', message_id=info_searching.id)
-            await asyncio.sleep(0.5)
-            await icbot.edit_message_text(
-                chat_id=msg.chat.id, text=f'**Downloading** {spotify.Sname}', message_id=info_searching.id)
-            await AlbumDownloader(inputs.text)
-            await icbot.delete_messages(msg.chat.id, info_searching.id)
-            await msg.reply_photo(photo=spotify.Scover, caption=f'Name : {spotify.Sname}\nArtists : {spotify.Sartist}', reply_markup=ReplyKeyboardRemove(True))
-            await msg.reply_document(document=spotify.Sname+'.zip', caption=f'🎧 Downloaded By [{Configs.BOT_NAME}](https://telegram.me/{Configs.BOT_UNAME}) Bot', reply_markup=MAIN_KEYS)
-            os.remove(spotify.Sname+'.zip')
+            result = spotify.Finder(0, inputs.text)
+            await icbot.edit_message_text(chat_id=msg.chat.id, text=spotify.Sname, message_id=info_searching.id)
+            if result:
+                await asyncio.sleep(0.5)
+                await icbot.edit_message_text(
+                    chat_id=msg.chat.id, text=f'**Downloading** {spotify.Sname}', message_id=info_searching.id)
+                await AlbumDownloader(inputs.text)
+                await icbot.delete_messages(msg.chat.id, info_searching.id)
+                await msg.reply_photo(photo=spotify.Scover, caption=f'Name : {spotify.Sname}\nArtists : {spotify.Sartist}', reply_markup=ReplyKeyboardRemove(True))
+                await msg.reply_document(document=spotify.Sname+'.zip', caption=f'🎧 Downloaded By [{Configs.BOT_NAME}](https://telegram.me/{Configs.BOT_UNAME}) Bot', reply_markup=MAIN_KEYS)
+                os.remove(spotify.Sname+'.zip')
         else:
             await msg.reply_text(f'You\'ve backed to main.', reply_markup=MAIN_KEYS)
 icbot.run()
